requirements.txt
accelerate>=1.5.0
Pillow>=10.4.0
sentencepiece>=0.2.0
transformers>=4.57.0
torchvision>=0.21.0
